requirements.txt
Flask==1.1.1
pymongo==3.10.1
python-dotenv==0.12.0
